requirements.txt
andas==3.0.1
psycopg2-binary==2.9.11